Removed an image optimization that happened in a rare situation
Before: An image whose visible pixels (if any) are all grayscale, but
which has some fully transparent pixels whose underlying color is not
grayscale, can be eligible to be written to a grayscale PNG image. This
results in grayscale-like garbage being written to the invisible parts
of the image. That's not *wrong*, but it's a bit strange.

After: Such images will not be optimized to grayscale, and the
underlying colors of invisible pixels should survive.